Unified ModeMap Type (#330)
This PR renames the type `OccupiedModeMap` to `ModeMap`.

## New features
* `occupied_mode_map` and `unoccupied_mode_map`— new constructors for
type `ModeMap`.
* `FermiUnoccupiedModes`— new iterator version generator of type
`ModeMap` for `FermiFS`.

## Deprecated features
* `OccupiedModeMap` is deprecated and replaced by `occupied_mode_map`.
The type is replaced by `ModeMap`.

Originally, there was a `OccupiedModeMap` type along with its
constructor function, which was used to generate a statically allocated
array for storing `*FSIndex` values corresponding to various Fock space
addresses. However, in some cases, it is also necessary to represent and
store unoccupied modes—particularly when working with Fermi addresses,
where the occupation number (occnum) is limited to 0 or 1.
The constructors enable consistent handling of both occupied and
unoccupied modes. Then the original `OccupiedModeMap` constructor
function is deprecated with `Base.@deprecate`.

---------

Co-authored-by: Joachim Brand <joachim.brand@gmail.com>
Co-authored-by: mtsch <matijacufar@gmail.com>