ModifiedHamiltonian (#331)
# New features
* Interface for creating Hamiltonians that modify the (off)-diagonal
elements with less boilerplate code. This is done by making the
Hamiltonian wrapper a `ModifiedHamiltonian` and implementing a small
interface. Implementations for `operator_column`, `offdiagonals`,
`diagonal_element` and a few other functions are then provided
automatically.
* Reimplementations of existing Hamiltonian wrappers (importance
sampling, symmetries, `FirstOrderTransitionOperator`, `TransformUndoer`,
`Stoquastic`) using the new interface. This makes them all work
correctly with Hamiltonians implemented using the new Hamiltonians
interface. Previously, that functionality was broken because the
wrappers were modifying the `offdiagonals` directly.
* New interface function `undo_transformation` that replaces the ad-hoc
way of handling transformations with `TransformUndoer`. The
transformations are now applied automatically. The results are reported
in the same columns as before. Passing `transform` to `AllOverlaps` is
now derpecated.
* New `IdentityOperator` which is currently only used in
`TransformUndoer`.
# Breaking changes
* The reimplemented wrappers no longer support the old Hamiltonians
interface.

---------

Co-authored-by: Joachim Brand <joachim.brand@gmail.com>
Co-authored-by: jamie-tay <jamiet96@gmail.com>