New constructors for `AbstractDVec`s (#333)
## Changed behaviour
* When constructing `DVec`, `InitiatorDVec` or `PDVec` with an iterator
or list of pairs, the data is added using `deposit!` i.e. using the
appropriate rules for adding data. In particular, zero value elements
won't be stored and values for multiple entries is added correctly.

## Fixed bugs
* Previously the constructors for `DVec`, `InitiatorDVec` and `PDVec`
passed arguments to a constructor for `Dict`. When multiple
contributions to the same address were passed, as when constructing from
an `AbstractOperatorColumn`, only the last contribution would be counted
and contributions not properly accumulated. This PR fixes this problem.

## Other changes
* Documentation for `DVec`, `InitiatorDVec` and `PDVec was updated.